Long overdue updates

+++ clink.sh
@@ -1,14 +1,21 @@
 #!/bin/sh
 
-if [ $# -eq 0 ]        
-  then                 
-    echo "No arguments supplied"               
-    exit 1             
-fi                     
+if [ $# -eq 0 ]
+then
+	echo "No arguments supplied"
+	exit 1
+fi
 
-FIRST_ARGUMENT="$1"    
+FIRST_ARGUMENT="$1"
 
-mkdir ~/.bak
-cp ~/$FIRST_ARGUMENT ~/.bak
-rm ~/$FIRST_ARGUMENT
+# only create folder if it doesn't exist
+mkdir -p ~/.bak
+
+# if file already exists, move it .bak
+if [ -e  ~/$FIRST_ARGUMENT ]
+then
+	cp ~/$FIRST_ARGUMENT ~/.bak
+	rm ~/$FIRST_ARGUMENT
+fi
+
 ln -s $(pwd)/$FIRST_ARGUMENT ~/$FIRST_ARGUMENT
